+++ src/cryptography/hazmat/primitives/kdf/kbkdf.py
@@ -28,6 +28,7 @@ class Mode(utils.Enum):
 class CounterLocation(utils.Enum):
     BeforeFixed = "before_fixed"
     AfterFixed = "after_fixed"
+    MiddleFixed = "middle_fixed"
 
 
 class _KBKDFDeriver:
@@ -39,6 +40,7 @@ class _KBKDFDeriver:
         rlen: int,
         llen: typing.Optional[int],
         location: CounterLocation,
+        break_location: typing.Optional[int],
         label: typing.Optional[bytes],
         context: typing.Optional[bytes],
         fixed: typing.Optional[bytes],
@@ -50,6 +52,24 @@ class _KBKDFDeriver:
 
         if not isinstance(location, CounterLocation):
             raise TypeError("location must be of type CounterLocation")
+
+        if break_location is None and location is CounterLocation.MiddleFixed:
+            raise ValueError("Please specify a break_location")
+
+        if (
+            break_location is not None
+            and location != CounterLocation.MiddleFixed
+        ):
+            raise ValueError(
+                "break_location is ignored when location is not"
+                " CounterLocation.MiddleFixed"
+            )
+
+        if break_location is not None and not isinstance(break_location, int):
+            raise TypeError("break_location must be an integer")
+
+        if break_location is not None and break_location < 0:
+            raise ValueError("break_location must be a positive integer")
 
         if (label or context) and fixed:
             raise ValueError(
@@ -79,6 +99,7 @@ class _KBKDFDeriver:
         self._rlen = rlen
         self._llen = llen
         self._location = location
+        self._break_location = break_location
         self._label = label
         self._context = context
         self._used = False
@@ -114,17 +135,29 @@ class _KBKDFDeriver:
         if rounds > pow(2, len(r_bin) * 8) - 1:
             raise ValueError("There are too many iterations.")
 
+        fixed = self._generate_fixed_input()
+
+        if self._location == CounterLocation.BeforeFixed:
+            data_before_ctr = b""
+            data_after_ctr = fixed
+        elif self._location == CounterLocation.AfterFixed:
+            data_before_ctr = fixed
+            data_after_ctr = b""
+        else:
+            if isinstance(
+                self._break_location, int
+            ) and self._break_location > len(fixed):
+                raise ValueError("break_location offset > len(fixed)")
+            data_before_ctr = fixed[: self._break_location]
+            data_after_ctr = fixed[self._break_location :]
+
         for i in range(1, rounds + 1):
             h = self._prf(key_material)
 
             counter = utils.int_to_bytes(i, self._rlen)
-            if self._location == CounterLocation.BeforeFixed:
-                h.update(counter)
+            input_data = data_before_ctr + counter + data_after_ctr
 
-            h.update(self._generate_fixed_input())
-
-            if self._location == CounterLocation.AfterFixed:
-                h.update(counter)
+            h.update(input_data)
 
             output.append(h.finalize())
 
@@ -152,6 +185,8 @@ class KBKDFHMAC(KeyDerivationFunction):
         context: typing.Optional[bytes],
         fixed: typing.Optional[bytes],
         backend: typing.Any = None,
+        *,
+        break_location: typing.Optional[int] = None,
     ):
         if not isinstance(algorithm, hashes.HashAlgorithm):
             raise UnsupportedAlgorithm(
@@ -178,6 +213,7 @@ class KBKDFHMAC(KeyDerivationFunction):
             rlen,
             llen,
             location,
+            break_location,
             label,
             context,
             fixed,
@@ -207,6 +243,8 @@ class KBKDFCMAC(KeyDerivationFunction):
         context: typing.Optional[bytes],
         fixed: typing.Optional[bytes],
         backend: typing.Any = None,
+        *,
+        break_location: typing.Optional[int] = None,
     ):
         if not issubclass(
             algorithm, ciphers.BlockCipherAlgorithm
@@ -226,6 +264,7 @@ class KBKDFCMAC(KeyDerivationFunction):
             rlen,
             llen,
             location,
+            break_location,
             label,
             context,
             fixed,